일정 리스트 구현

+++ BootHack/src/main/java/com/example/demo/repository/MemberRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.MemberModel;
 
 @Repository
-public interface MemberRepository extends JpaRepository<MemberModel, Long> {
+public interface MemberRepository extends JpaRepository<MemberModel, Integer> {
 	
 	MemberModel findByIdAndPw(String id, String pw);
 }
